Move state from SearchResults to WebPageRoutes so data can be used everywhere
