+++ src/containers/Homepage/index.js
@@ -3,11 +3,14 @@ import React from "react";
 import { CuisineListContainer, LocationSearchForm } from "../index";
 
 const Homepage = (props) => {
-  const { setUserInput } = props;
+  const { setUserInput, handleFetch } = props;
 
   return (
     <>
-      <LocationSearchForm setUserInput={setUserInput} />
+      <LocationSearchForm
+        setUserInput={setUserInput}
+        handleFetch={handleFetch}
+      />
       <CuisineListContainer />
     </>
   );
