+++ src/containers/SearchResults/index.js
@@ -1,39 +1,37 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 
-import { filterByRating } from "../../utils/filterByRating";
 import { LoadingConfirmation, RestaurantsNear } from "../../components";
 import { RestaurantList } from "../index";
-import { yelp } from "../../utils/fetchRestaurantData";
 
 const SearchResults = (props) => {
-  const { userInput } = props;
-  const [restaurantData, updateRestaurantData] = useState([]);
-  const [isLoading, updateLoadingState] = useState(false);
-  const [defaultRestaurantOrder, applyDefaultRestaurantOrder] = useState([]);
+  const { userInput, isLoading, data, handleOnChange } = props;
+  // const [restaurantData, updateRestaurantData] = useState([]);
+  // const [isLoading, updateLoadingState] = useState(false);
+  // const [defaultRestaurantOrder, applyDefaultRestaurantOrder] = useState([]);
 
-  useEffect(() => {
-    yelp.getData(userInput, updateRestaurantData, updateLoadingState);
-    // updateLoadingState only works within the method call - need to look into
-  }, [userInput]);
+  // useEffect(() => {
+  //   yelp.getData(userInput, updateRestaurantData, updateLoadingState);
+  //   // updateLoadingState only works within the method call - need to look into
+  // }, [userInput]);
 
-  const handleOnChange = (action) => {
-    if (!defaultRestaurantOrder.length) {
-      applyDefaultRestaurantOrder([...restaurantData]);
-    }
-    const actionList = {
-      rating: filterByRating(restaurantData),
-      nearest: [...defaultRestaurantOrder],
-    };
+  // const handleOnChange = (action) => {
+  //   if (!defaultRestaurantOrder.length) {
+  //     applyDefaultRestaurantOrder([...restaurantData]);
+  //   }
+  //   const actionList = {
+  //     rating: filterByRating(restaurantData),
+  //     nearest: [...defaultRestaurantOrder],
+  //   };
 
-    updateRestaurantData(actionList[action]);
-  };
+  //   updateRestaurantData(actionList[action]);
+  // };
 
   return (
     <>
       <RestaurantsNear userInput={userInput} />
       {isLoading && <LoadingConfirmation />}
       <RestaurantList
-        data={restaurantData}
+        data={data}
         handleOnChange={handleOnChange}
         isLoading={isLoading}
       />
